Put access of overall attributes behind setters, too
