+++ result/overall_test.go
@@ -10,14 +10,6 @@ import (
 	"github.com/NETWAYS/go-check"
 )
 
-func TestOverall_NewPartialResult(t *testing.T) {
-	actual := NewPartialResult()
-
-	if actual.String() != "[UNKNOWN] " {
-		t.Fatalf("expected '[UNKNOWN] ', got %s", actual.String())
-	}
-}
-
 func TestOverall_AddOK(t *testing.T) {
 	overall := Overall{}
 	overall.Add(0, "test ok")
@@ -108,7 +100,7 @@ func TestOverall_GetOutput(t *testing.T) {
 	overall = Overall{}
 	overall.Add(0, "First OK")
 	overall.Add(2, "Second Critical")
-	overall.OKSummary = "Custom Summary"
+	overall.oKSummary = "Custom Summary"
 
 	expected = "Second Critical\n\\_ [OK] First OK\n\\_ [CRITICAL] Second Critical\n"
 
@@ -171,7 +163,7 @@ func ExampleOverall_withSubchecks() {
 func ExampleOverall_withVerticalbar() {
 	var overall Overall
 
-	overall.OKSummary = "unit|test"
+	overall.oKSummary = "unit|test"
 
 	subcheck := &PartialResult{
 		output: "vertical|bar",
@@ -583,7 +575,7 @@ func TestOverallGetOutput_WithMultipleStatesMultipleTimes(t *testing.T) {
 }
 
 func TestOverall_Add_WithRace(t *testing.T) {
-	o := &Overall{OKSummary: "unittest"}
+	o := &Overall{oKSummary: "unittest"}
 
 	var wg sync.WaitGroup
 
@@ -616,7 +608,7 @@ func TestOverall_AddSubcheck_WithRace(t *testing.T) {
 }
 
 func TestOverall_Get_WithRace(t *testing.T) {
-	o := &Overall{OKSummary: "unittest"}
+	o := &Overall{oKSummary: "unittest"}
 
 	for range 3 {
 		o.Add(check.OK, "OK")
@@ -640,59 +632,5 @@ func TestOverall_Get_WithRace(t *testing.T) {
 		}()
 	}
 
-	wg.Wait()
-}
-
-func TestPartialResult_SetGet_WithRace(t *testing.T) {
-	pr := NewPartialResult()
-
-	var wg sync.WaitGroup
-
-	for range 3 {
-		wg.Add(2)
-		go func() {
-			defer wg.Done()
-			pr.SetState(check.Critical)
-		}()
-		go func() {
-			defer wg.Done()
-			_ = pr.GetStatus()
-		}()
-	}
-	wg.Wait()
-}
-
-func TestPartialResult_AddSubcheck_WithRace(t *testing.T) {
-	parent := NewPartialResult()
-	parent.SetOutput("unittest")
-
-	var wg sync.WaitGroup
-	for range 5 {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			child := NewPartialResult()
-			child.SetState(check.OK)
-			parent.AddSubcheck(child)
-		}()
-	}
-	wg.Wait()
-}
-
-func TestPartialResult_SetDefaultState_WithRace(t *testing.T) {
-	pr := NewPartialResult()
-
-	var wg sync.WaitGroup
-	for range 5 {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-
-			example_perfdata := check.Perfdata{Label: "pd_test", Value: 5, Uom: "s"}
-			pr.AddPerfdata(&example_perfdata)
-
-			pr.SetDefaultState(check.Warning)
-		}()
-	}
 	wg.Wait()
 }
